+++ cpp/subprojects/boosting/src/mlrl/boosting/rule_evaluation/rule_evaluation_decomposable_complete_binned.cpp
@@ -5,7 +5,7 @@
 namespace boosting {
 
     DecomposableCompleteBinnedRuleEvaluationFactory::DecomposableCompleteBinnedRuleEvaluationFactory(
-      float64 l1RegularizationWeight, float64 l2RegularizationWeight,
+      float32 l1RegularizationWeight, float32 l2RegularizationWeight,
       std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr)
         : l1RegularizationWeight_(l1RegularizationWeight), l2RegularizationWeight_(l2RegularizationWeight),
           labelBinningFactoryPtr_(std::move(labelBinningFactoryPtr)) {}
